+++ main/trans/old_model/shop_user.go
@@ -9,7 +9,7 @@ import (
 )
 
 type ShopUser struct {
-	ShopUserID       uint   `gorm:"primaryKey;autoIncrement;comment:主键id"`
+	ShopUserID       uint64 `gorm:"primaryKey;autoIncrement;comment:主键id"`
 	UserName         string `gorm:"default:'';comment:用户名"`
 	Password         string `gorm:"default:'';comment:登录密码"`
 	Phone            string `gorm:"default:'';comment:手机号"`
@@ -20,13 +20,13 @@ type ShopUser struct {
 	IsDelete         int64  `gorm:"default:0;comment:0=显示1=伪删除"`
 	UserType         uint   `gorm:"default:0;comment:账号类型0总台1门店"`
 	IsStatus         uint   `gorm:"default:0;comment:是否禁用1禁用，0未禁用"`
-	AppID            uint   `gorm:"default:0;comment:程序id"`
+	AppID            uint64 `gorm:"default:0;comment:程序id"`
 	BindKey          string `gorm:"default:'';comment:绑定的设备key"`
 	CashierOnline    uint   `gorm:"default:0;comment:收银员当班 0-不在线 1-在线"`
 	CashierLoginTime uint   `gorm:"default:0;comment:收银员当班登录时间"`
 	DutyNo           string `gorm:"default:'';comment:当班编号"`
-	CreateTime       int64  `gorm:"autoCreateTime;comment:创建时间"`
-	UpdateTime       int64  `gorm:"autoUpdateTime;comment:更新时间"`
+	CreateTime       int    `gorm:"autoCreateTime;comment:创建时间"`
+	UpdateTime       int    `gorm:"autoUpdateTime;comment:更新时间"`
 }
 
 type ShopUserRepository interface {
@@ -58,18 +58,18 @@ func (s *ShopUserService) ConvertShopUser() error {
 			Username:         shopUser.UserName,
 			Password:         shopUser.Password,
 			Phone:            shopUser.Phone,
-			PasswordChange:   shopUser.PasswordChange,
+			PasswordChange:   int(shopUser.PasswordChange),
 			RealName:         shopUser.RealName,
-			IsSuper:          shopUser.IsSuper,
-			UserType:         shopUser.UserType,
-			IsDisable:        shopUser.IsStatus,
+			IsSuper:          int(shopUser.IsSuper),
+			UserType:         int(shopUser.UserType),
+			IsDisable:        int(shopUser.IsStatus),
 			BindKey:          shopUser.BindKey,
-			CashierOnline:    shopUser.CashierOnline,
-			CashierLoginTime: shopUser.CashierLoginTime,
+			CashierOnline:    int(shopUser.CashierOnline),
+			CashierLoginTime: int(shopUser.CashierLoginTime),
 			DutyNo:           shopUser.DutyNo,
 			CreateTime:       shopUser.CreateTime,
 			UpdateTime:       shopUser.UpdateTime,
-			DeleteTime:       shopUser.IsDelete,
+			DeleteTime:       int(shopUser.IsDelete),
 		}
 		err := repository.NewStaffRepo(s.targetDB).CreateStaff(staff)
 		if err != nil {
